+++ All.ui/miniprogram/components/picker/region.js
@@ -1,6 +1,5 @@
 
 module.exports = {
-  
   list: [{
   "code": "110000",
   "name": "北京市",